feat(pr): use `PathRelinkingState` only when it's actually running
lol so the bug was that initially, `LocalSearchState` was designed for
both "regular" local search and Path Relinking local search,
but that became unnecessary when I started to split
the local search methods into regular and PR anyways,
so there was no point in sharing state between both.

This latent bug was discovered when I started to apply
a regular local search to relinked solutions during PR
since they're not guaranteed to be local optima.