feat(pr): store relinked path before applying A-FVS
there's no guarantee that a relinked solution
(a path found in Path Relinking)
is a local optimum, but running A-FVS (a "regular" local search)
on a relinked can change it
in such a way that it could no longer be a path between
the starting (S_s) and target (S_t) solutions,
causing a bug.

Now the relinked is saved so it can be restored after running A-FVS,
allowing to restore the relinked between S_s and S_t
and continue exploring the rest of the paths.